script/requirements.txt
aiohttp==3.6.2
alpaca-trade-api==0.49.0
alpha-vantage==2.2.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
finnhub-python==2.1.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.1
pandas==1.1.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
wrapt==1.12.1
yarl==1.5.1
